[FIO toup] arm: boot: reserve devices for exclusive access
Drivers defined in the platform specific main.c module will be added
to the dts and to the generated overlay (if overlay generation is
requested)

In order to do that, the platforms need to define an interface in
their main.c module through which the generic code can retrieve the
driver's name.

This implementation supports a case if a middle-way sub-node exists.

An example for such an interface in iMX platforms can be seen below
(each specific iMX platform will have a different set of requirements).

const char *main_get_optee_exclusive_node_name(unsigned int i)
{
	static const char *exclusive_drivers[] = {
	    	"/soc/aips-bus@2200000/crypto@2280000",
		NULL,
	};

	if (i < ARRAY_SIZE(exclusive_drivers))
		return exclusive_drivers[i];

	return NULL;
}

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>
Co-developed-by: Oleksandr Suvorov <oleksandr.suvorov@foundries.io>
Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@foundries.io>